ipo_ml_screener/src/ipo_ml_screener/config.py
from __future__ import annotations
from pathlib import Path
import yaml


def load_config(path: str | Path = "config.yaml") -> dict:
    p = Path(path)
    if not p.exists():
        raise FileNotFoundError(f"Config not found: {p.resolve()}")
    return yaml.safe_load(p.read_text(encoding="utf-8"))
