ipo_ml_screener/requirements.txt
streamlit>=1.34
pandas>=2.1
numpy>=1.26
requests>=2.31
yfinance>=0.2.40
pyyaml>=6.0.1
beautifulsoup4>=4.12
lxml>=5.2
plotly>=5.22
pyarrow>=16.1
